requirements.txt
boto3==1.42.54
pydantic==2.12.5
pytz==2025.2
python-dotenv==1.2.1
